+++ src/components/Cronograma.js
@@ -1,23 +1,24 @@
 import React from 'react'
 import { StyleSheet, Text, View, Button } from 'react-native'
+import PropTypes from 'prop-types'
 
-function Cronograma({ name, navigation }) {
-  // TODO: Editar and Trabajar could go to a i19n file
+function Cronograma({ id, name, navigation }) {
+  // TODO: Editar and Trabajar could go to a i18n file
   return (
     <View style={styles.container}>
       <View style={styles.rowContainer}>
-        <Text>{name}</Text>
-        <Button
-          title="Editar"
-          onPress={() => navigation.navigate('Editar')}
-        />
-        <Button
-          title="Ver"
-          onPress={() => navigation.navigate('Trabajar')}
-        />
+        <Text>{id} {name}</Text>
+        <Button title="Editar" onPress={() => navigation.navigate('Editar', { id })} />
+        <Button title="Ver" onPress={() => navigation.navigate('Trabajar', { id })} />
       </View>
     </View>
   )
+}
+
+Cronograma.propTypes = {
+  name: PropTypes.string.isRequired,
+  navigation: PropTypes.object.isRequired,
+  id: PropTypes.number.isRequired,
 }
 
 const styles = StyleSheet.create({