+++ src/screens/WorkView.js
@@ -1,16 +1,25 @@
-import { StyleSheet, Text, View, Button } from 'react-native'
 import React from 'react'
+import { Text, View, Button } from 'react-native'
+import PropTypes from 'prop-types'
 
-export default function WorkView({navigation}) {
+import CardList from '../components/CardList'
+
+function WorkView({navigation , route}) {
+
+  const { id } = route.params
+
   return (
     <View>
-      <Text>WorkView</Text>
-      <Button
-            title="Ayuda"
-            onPress={() => navigation.navigate('Informacion')}
-        />
+      <Text>Perfil ID: {id}</Text>
+      <CardList scheduleId={id} />
+      <Button title="Ayuda" onPress={() => navigation.navigate('Informacion')} />
     </View>
   )
 }
 
-const styles = StyleSheet.create({})
+WorkView.propTypes = {
+  navigation: PropTypes.object.isRequired,
+  route: PropTypes.object.isRequired,
+}
+
+export default WorkView